fix: rename wandb.py to mixins.py in an attempt to fix remote test errors
