+++ mirascope/wandb/__init__.py
@@ -1,3 +1,3 @@
 """Integrations with Weights & Biases (wandb, weave) tooling."""
-from .wandb import WandbCallMixin, WandbExtractorMixin, trace, trace_error
+from .mixins import WandbCallMixin, WandbExtractorMixin, trace, trace_error
 from .weave import with_weave
